+++ docs/vscode_cheat_debug.txt
@@ -1,0 +1,13 @@
+Continue (F5): Continues execution until the next breakpoint.
+Step Over (F10): Executes the next line of code but doesn’t step into any functions. It’s useful when you want to execute a function call but don’t want to debug inside the function.
+Step Into (F11): Steps into the function call. If the next line is a function call, it will take you inside the function.
+Step Out (Shift+F11): Steps out of the current function and returns to the caller function. It’s useful when you are done debugging inside a function and want to return to the higher level.
+Restart (Ctrl+Shift+F5): Restarts the debugging session.
+Stop (Shift+F5): Stops the debugging session.
+Using the Debugging Controls
+Continue (F5): Use this to resume the execution of your code until it hits the next breakpoint.
+Step Over (F10): If you’re at a line that calls a function, use Step Over to execute the function without diving into it.
+Step Into (F11): If you want to dive into the details of a function call, use Step Into.
+Step Out (Shift+F11): If you’ve stepped into a function and now want to return to the calling function, use Step Out.
+Restart (Ctrl+Shift+F5): If you need to restart your debugging session.
+Stop (Shift+F5): To stop the debugging session.
