+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+markers =
+    unit: mark a test as a unit test.
+    integration: mark a test as an integration test.
