+++ requirements-dev.txt
@@ -1,0 +1,2 @@
+debugpy
+pytest
