+++ scripts/install_python_dependencies.sh
@@ -8,6 +8,7 @@ source /workspace/venv/bin/activate
 
 # Install Python dependencies from requirements.txt
 pip install --no-cache-dir -r /workspace/requirements.txt
+pip install --no-cache-dir -r /workspace/requirements-dev.txt
 
 # List installed packages to verify installation
 pip list
