+++ src/pipeline.py
@@ -1,13 +1,15 @@
-# src/pipeline.py
-
 import os
 import logging
-from preprocess.preprocess_step import PreprocessStep
-from ocr.ocr_step import OCRStep
+from step_01_preprocess.preprocess_step import PreprocessStep
+from step_02_ocr.ocr_step import OCRStep
 
 # Constants
 INPUT_DIRECTORY = '/workspace/data'
 LOG_FILE = '/workspace/data/pipeline.log'
+PATH_TO_TESSERACT= '/usr/share/tesseract-ocr/4.00/tessdata'
+
+# Add PATH_TO_TESSERACT to args
+args.path_to_tesseract = PATH_TO_TESSERACT
 
 # Setup logging
 logging.basicConfig(filename=LOG_FILE, level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
@@ -16,17 +18,17 @@ def run_pipeline(args):
     logging.info("Starting pipeline execution")
     
     # Step 1: Preprocessing
-    preprocess_step = PreprocessStep(args)
-    preprocess_step.run(INPUT_DIRECTORY)
-
-    # Allow user intervention here if needed
-    input("Press Enter to continue to OCR step...")
+    step_01 = PreprocessStep(args)
+    step_01.run(INPUT_DIRECTORY)
 
     # Step 2: OCR
-    ocr_step = OCRStep(args.language, '/usr/share/tesseract-ocr/4.00/tessdata', args.check_orientation, args.psm)
-    ocr_step.run(os.path.join(INPUT_DIRECTORY, 'preprocessed'))
+    step_02 = OCRStep(PATH_TO_TESSERACT, args)
+    step_02.run(os.path.join(INPUT_DIRECTORY, 'preprocessed'))
 
     logging.info("Pipeline execution completed successfully")
+
+    # Step 3: ...
+    
 
 if __name__ == "__main__":
     import argparse
@@ -38,7 +40,6 @@ if __name__ == "__main__":
     parser.add_argument('--erode', action='store_true', help='Apply erosion')
     parser.add_argument('--opening', action='store_true', help='Apply opening (erosion followed by dilation)')
     parser.add_argument('--canny', action='store_true', help='Apply Canny edge detection')
-    parser.add_argument('--deskew', action='store_true', help='Apply deskewing (skew correction)')
     parser.add_argument('--language', type=str, default='eng', help='Language for Tesseract OCR')
     parser.add_argument('--check-orientation', type=int, choices=[0, 1], default=0, help='Check and correct orientation')
     parser.add_argument('--psm', type=int, choices=list(range(14)), default=6, help='Tesseract Page Segmentation Mode (PSM)')
