+++ tests/test_pipeline.py
@@ -1,9 +1,9 @@
-# tests/test_pipeline.py
-
-import pytest
 import os
-from src.pipeline import run_pipeline
+import pytest
+from PIL import Image
+import numpy as np
 from argparse import Namespace
+from src.pipeline import run_pipeline
 
 def test_pipeline():
     args = Namespace(
@@ -17,12 +17,13 @@ def test_pipeline():
         deskew=False,
         language='eng',
         check_orientation=True,
-        psm=6
+        psm=6,
+        no_pause=True
     )
 
     input_data = "/tmp/test_pipeline_data"
     os.makedirs(input_data, exist_ok=True)
-    
+
     # Create dummy images
     dummy_image_path = os.path.join(input_data, 'test_image.png')
     image = Image.fromarray(np.zeros((100, 100), dtype=np.uint8))
